# backend/docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WatchRuleCreate": {
        "example": {
          "is_active": true,
          "name": "Rare techno under $40",
          "poll_interval_seconds": 600,
          "query": {
            "max_price": 40,
            "q": "detroit techno",
            "sources": [
              "discogs"
            ]
          }
        },
        "properties": {
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "poll_interval_seconds": {
            "default": 600,
            "maximum": 86400.0,
            "minimum": 30.0,
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          }
        },
        "required": [
          "name"
        ],
        "title": "WatchRuleCreate",
        "type": "object"
      },
      "WatchRuleOut": {
        "example": {
          "created_at": "2026-01-20T11:52:00+00:00",
          "id": "80dc6333-3c3c-49b8-a803-938783fbeb99",
          "is_active": true,
          "last_run_at": "2026-01-20T12:00:00+00:00",
          "name": "Rare techno under $40",
          "next_run_at": "2026-01-20T12:10:00+00:00",
          "poll_interval_seconds": 600,
          "query": {
            "max_price": 40,
            "q": "detroit techno",
            "sources": [
              "discogs"
            ]
          },
          "updated_at": "2026-01-20T12:00:00+00:00",
          "user_id": "8f2a5009-c0a2-4f90-8f1b-c1716c26bf06"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          },
          "poll_interval_seconds": {
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "query",
          "is_active",
          "poll_interval_seconds",
          "last_run_at",
          "next_run_at",
          "created_at",
          "updated_at"
        ],
        "title": "WatchRuleOut",
        "type": "object"
      },
      "WatchRuleUpdate": {
        "example": {
          "is_active": true,
          "name": "Rare techno under $30",
          "poll_interval_seconds": 900,
          "query": {
            "max_price": 30,
            "q": "detroit techno",
            "sources": [
              "discogs"
            ]
          }
        },
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "poll_interval_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 30.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll Interval Seconds"
          },
          "query": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          }
        },
        "title": "WatchRuleUpdate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "watch-wax-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/watch-rules": {
      "get": {
        "operationId": "list_rules_api_watch_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchRuleOut"
                  },
                  "title": "Response List Rules Api Watch Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Rules",
        "tags": [
          "watch-rules"
        ]
      },
      "post": {
        "operationId": "create_rule_api_watch_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Rule",
        "tags": [
          "watch-rules"
        ]
      }
    },
    "/api/watch-rules/{rule_id}": {
      "delete": {
        "operationId": "disable_rule_api_watch_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable Rule",
        "tags": [
          "watch-rules"
        ]
      },
      "get": {
        "operationId": "get_rule_api_watch_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Rule",
        "tags": [
          "watch-rules"
        ]
      },
      "patch": {
        "operationId": "patch_rule_api_watch_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchRuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Rule",
        "tags": [
          "watch-rules"
        ]
      }
    },
    "/api/watch-rules/{rule_id}/disable": {
      "post": {
        "operationId": "disable_rule_explicit_api_watch_rules__rule_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable Rule Explicit",
        "tags": [
          "watch-rules"
        ]
      }
    },
    "/api/watch-rules/{rule_id}/hard": {
      "delete": {
        "operationId": "delete_rule_hard_api_watch_rules__rule_id__hard_delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Rule Hard",
        "tags": [
          "watch-rules"
        ]
      }
    },
...
  }
}